+++ nubix/__init__.py
@@ -1,5 +1,5 @@
 """Nubix - Cloud Sync Manager for Ubuntu."""
 
-__version__ = "4.3.0"
+__version__ = "4.4.0"
 __author__ = "Nubix Contributors"
 __license__ = "Proprietary"